+++ modules/ccv2/core/src/sap/commerce/toolset/ccv2/event/CCv2Listener.kt
@@ -19,12 +19,15 @@
 package sap.commerce.toolset.ccv2.event
 
 import sap.commerce.toolset.ccv2.dto.CCv2Dto
+import sap.commerce.toolset.ccv2.dto.CCv2EnvironmentDto
 import sap.commerce.toolset.ccv2.settings.state.CCv2Subscription
 
 sealed interface CCv2Listener<T : CCv2Dto> {
-    fun onFetchingStarted(subscriptions: Collection<CCv2Subscription>) = Unit
+    fun onFetchingStarted(data: Collection<CCv2Subscription>) = Unit
     fun onFetchingCompleted(data: Map<CCv2Subscription, Collection<T>> = emptyMap()) = Unit
 
     fun onFetchingBuildDetailsStarted(data: Map<CCv2Subscription, Collection<T>>) = Unit
     fun onFetchingBuildDetailsCompleted(data: Map<CCv2Subscription, Collection<T>> = emptyMap()) = Unit
+
+    fun onEndpointUpdate(data: CCv2EnvironmentDto) = Unit
 }