+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## [2025.2.4.4]
 
 <cite>Release contributors</cite>
-- 1 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.4.5+author%3Amlytvyn+is%3Apr)
+- 3 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.4.5+author%3Amlytvyn+is%3Apr)
 
 ### `CCv2` enhancements
 - Display available endpoints per environment [#1616](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1616)
 - Display state of the `maintenance mode` of an endpoint [#1617](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1617)
+- Added the ability to toggle `maintenance mode` for an endpoint [#1618](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1618)
 
 ## [2025.2.4.4]